start debug session
