+++ src/solver/optimisation/opt_construction_matrice_des_contraintes_cas_lineaire.cpp
@@ -47,6 +47,7 @@
 #include "AreaHydroLevel.h"
 #include "FinalStockEquivalent.h"
 #include "FinalStockExpression.h"
+#include "constraint_builder_utils.h"
 
 #include <antares/study.h>
